+++ src/enums/index.ts
@@ -1,4 +1,5 @@
 export { ROUTE_NAMES } from './route-names.enum'
 export { ICON_NAMES } from './icon-names.enum'
 export { REQUEST_STATUSES } from '@/enums/statuses.enum'
+export { MAX_LENGTH } from './validation.emum'
 export * from './window-breakpoints.enum'